Added `subpage` event for creating subpages
The subpage event takes the current route and adds a `#` character
after it together with a new route. The crossroads parser will only
see the new route, and the rest of the route will be used for
breadcrumbs display.

There is a new variable in router.js called routesPath. It contains
the following properties:

* literal - a literal route, for example: route1/john#route2/ann
* abstract - an abstract route, for example: route1/{name}#route2/{wife}

The subpage can also be used with knockout's data binding, like this:

    <a class="link" data-bind="subpage: 'route2/' + id"></a>

and this will activate the `subpage` event on click.